requirements.txt
fastapi
uvicorn
pydantic[email]
boto3
python-multipart
Pillow
pdf2image
ffmpeg-python
python-jose[cryptography]
passlib[bcrypt]
elasticsearch>=8.0.0

SQLAlchemy

asyncpg
databases

